Changement positionnement fichiers.
