+++ ModulesCoeur/Demarrage.py
@@ -1,8 +1,7 @@
-import imp
 import os
 from ModulesCoeur.Initialisation import *
 from ClassesCoeur.UserInterface import *
-from ClassesExt.Comprendre import *
+from ModulesCoeur.Comprendre import *
 
 
 def verifierInit(USER_NAME, USERInterface, chemin_DATA, chemin_user):
